code/src/requirements.txt
Flask==2.2.3
python-dotenv==0.21.0
openai==0.27.0
pdfplumber==0.8.0
pytesseract==0.3.8
python-docx==0.8.11
pywin32==310